Shorter version of video owner
using populate()

+++ src/controllers/videoController.js
@@ -19,19 +19,29 @@ export const watch = async(req, res) => {
 
 export const getEdit = async(req, res) =>  {
     const { id } = req.params;
+    const { user: {_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.render("404", {pageTitle: "404 Error, Video not found"});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     return res.render("edit", {pageTitle: `Editing ${video.title}`, video});
 };
 
 export const postEdit = async(req, res) => {
+    const {
+        user: {_id}
+    } = req.session;
     const { id } = req.params;
     const {title, description, hashtags} = req.body;
     const video = await Video.exists({_id:id});
     if(!video){
         return res.render("404", {pageTitle: "404 Error, Video not found"});
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -50,13 +60,16 @@ export const postUpload = async (req, res) => {
     const {path:fileUrl} = req.file;
     const { title, description, hashtags } = req.body;
     try{
-        await Video.create({
+        const newVideo = await Video.create({
             title, //abbreviation of title: title, 
             description,
             fileUrl,
             owner: _id,
             hashtags: Video.formatHashtags(hashtags),
         });
+        const user = await User.findById(_id);
+        user.videos.push(newVideo._id);
+        user.save();
         return res.redirect("/");
     }
     catch(error){
@@ -68,8 +81,17 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async(req, res) => {
+    const {
+        user: {_id}
+    } = req.session;
     const { id } = req.params;
-    console.log(id);
+    const video = await Video.findById(id);
+    if(!video){
+        return res.render("404", { pageTitle: "404 Error, Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     // delete video
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
